requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio
matplotlib
numpy
pandas
scikit-learn
opencv-python
albumentations
scipy
networkx
diffusers>=0.20.0
transformers
grad-cam
tqdm
tensorboard
jupyter
Pillow
gdown
